refactor(BA-6667): move status enums down from models to data
Move NetworkType into data/network/types.py and redirect the data-layer
imports of SessionStatus/KernelStatus (already defined in data) so
data/sokovan/* and data/session/* stop importing upward into models.
This removes 7 of the data->models back-edges feeding the models<->data
cycle. models/network/row.py re-exports NetworkType for backward compat.
