+++ src/ai/backend/manager/data/sokovan/result.py
@@ -5,10 +5,9 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 
 from ai.backend.common.types import KernelId, SessionId
-from ai.backend.manager.data.session.types import KernelMatchType
+from ai.backend.manager.data.kernel.types import KernelStatus
+from ai.backend.manager.data.session.types import KernelMatchType, SessionStatus
 from ai.backend.manager.exceptions import ErrorStatusInfo
-from ai.backend.manager.models.kernel import KernelStatus
-from ai.backend.manager.models.session import SessionStatus
 
 from .lifecycle import SessionWithKernels
