+++ src/ai/backend/manager/data/sokovan/workload.py
@@ -15,7 +15,7 @@ from ai.backend.common.types import (
     SessionResult,
     SessionTypes,
 )
-from ai.backend.manager.models.session import SessionStatus
+from ai.backend.manager.data.session.types import SessionStatus
 
 
 @dataclass(frozen=True)
